完善和优化商城，目前支持淘宝、京东、拼多多等平台的查询，详情、推广等。优化订单管理部分

+++ src/main/java/com/cn/leedane/controller/BlogController.java
@@ -62,7 +62,7 @@ public class BlogController extends BaseController{
 		 */
 		boolean hasImg = JsonUtil.getBooleanValue(json, "has_img");
 
-		int blogId = JsonUtil.getIntValue(json, "bid");
+		long blogId = JsonUtil.getLongValue(json, "bid");
 		
 		/**
 		 * 是否要自动截取摘要摘要
