+++ src/main/java/com/cn/leedane/model/FanBean.java
@@ -12,15 +12,15 @@ public class FanBean extends RecordTimeBean{
 
 	//状态：1正常
 	private static final long serialVersionUID = 1L;
-	private int toUserId; //接收粉丝的用户ID
+	private long toUserId; //接收粉丝的用户ID
 	
 	private String userRemark;  //FromUserId对toUserId对应的备注信息
 	
 	//@Column(name="to_user_id")
-	public int getToUserId() {
+	public long getToUserId() {
 		return toUserId;
 	}
-	public void setToUserId(int toUserId) {
+	public void setToUserId(long toUserId) {
 		this.toUserId = toUserId;
 	}
 	//@Column(name="user_remark")
