+++ src/main/java/com/cn/leedane/model/PrivateChatBean.java
@@ -15,17 +15,17 @@ public class PrivateChatBean extends RecordTimeBean{
 	 */
 	private static final long serialVersionUID = 4673215292460039597L;
 
-	private int toUserId; //接收好友的用户ID
+	private long toUserId; //接收好友的用户ID
 	
 	private String content ;  //发送消息的内容
 	
 	private int type; //发送消息的类型，0：文本,1:语音, 3:图片
 	
 	//@Column(name="to_user_id")
-	public int getToUserId() {
+	public long getToUserId() {
 		return toUserId;
 	}
-	public void setToUserId(int toUserId) {
+	public void setToUserId(long toUserId) {
 		this.toUserId = toUserId;
 	}
 	public String getContent() {
